test_feedTools.py
import pytest


# TESTING CANNOT BE DONE ON GITHUB BECAUSE IT REQUIRES KEYS, SO IT HAS BEEN COMMENTED OUT

# # Testing that getFeed() actually obtains a twitter users feed. CAnnot be nore specific than non-empty return value
# # because the twitter feed of every user is different
# def test_successfeed():
#     assert len(feedTools.getFeed()) > 0


# # Testing that the annotateImage() function is correctly annotating images with Googles Vision API.
# def test_imageVision():
#     assert feedTools.annotateImage(
#         "https://pbs.twimg.com/media/EQTfQDoUwAIMsRY?format=jpg&name=4096x4096") == "Basketball player, Basketball, Basketball moves, Player, Sports, Team sport, Basketball court, Ball game, Tournament, Muscle"


def test_true():
    assert True == True
